factor Byron client/server from Main
